+++ src/core/alerts/clear-seearch-history-alert.ts
@@ -1,0 +1,16 @@
+import { Alert } from 'react-native'
+
+export const clearSearchHistoryAlert = (
+  onPress: () => void,
+) => {
+  Alert.alert('Are you sure?', 'Do you really want to clear your search history?', [
+    {
+      text: 'Cancel',
+      onPress: () => null,
+    },
+    {
+      text: 'Clear',
+      onPress,
+    }
+  ])
+}