+++ src/core/hooks/use-rehydrate-store.hook.ts
@@ -4,15 +4,27 @@ import { useDispatch } from 'react-redux'
 import { Storage } from '../storage/storage'
 import { STORAGE_KEY } from '../storage/storage.constants'
 
-export const useRehydrateStore = (
+interface UseRehydrateStore<V> {
+  value: V | null;
+  isPending: boolean;
+}
+
+export const useRehydrateStore = <V>(
   key: STORAGE_KEY,
-  action: (value: any) => AnyAction,
-) => {
+  action: (value: V) => AnyAction,
+): UseRehydrateStore<V> => {
+  const [ value, setValue ] = React.useState<V | null>(null)
+  const [ isPending, setPending ] = React.useState(true)
+
   const dispatch = useDispatch()
 
   React.useEffect(() => {
     Storage.get(key).then((value: any) => {
+      setValue(value)
+      setPending(false)
       dispatch(action(value))
     })
   }, [])
+
+  return { value, isPending }
 }