+++ src/more/more.tsx
@@ -2,18 +2,27 @@ import * as React from 'react'
 import * as R from 'ramda'
 import app from '../../package.json'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
+import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons'
 import { MoreActivityIndicatorWrapper, MoreContainer } from './more.styled'
 import { ListedOption } from './components/listed-option/listed-option'
 import { CustomSwitch } from '../core/custom-switch/custom-switch'
 import { TEXT_SIZE } from '../core/text/text.constants'
-import { ActivityIndicator, FlatList } from 'react-native'
+import { ActivityIndicator, FlatList, Platform, TouchableOpacity } from 'react-native'
 import { useDispatch, useSelector } from 'react-redux'
-import { hapticFeedbackEnabledSelector } from '../settings/store/settings.selectors'
-import { setHapticFeedbackEnabledAction } from '../settings/store/settings.slice'
+import { hapticFeedbackEnabledSelector, nativeSearchEngineEnabledSelector } from '../settings/store/settings.selectors'
+import { setHapticFeedbackEnabledAction, setNativeSearchEngineEnabledAction } from '../settings/store/settings.slice'
+import { useRehydrateStore } from '../core/hooks/use-rehydrate-store.hook'
+import { STORAGE_KEY } from '../core/storage/storage.constants'
+import { Storage } from '../core/storage/storage'
+import { CustomButton } from '../core/custom-button/custom-button'
+import { COLOR } from '../core/colors/colors.constants'
+import { clearSearchHistoryAlert } from '../core/alerts/clear-seearch-history-alert'
 
 interface MoreOption<V> {
   title: string;
   onChange: (newValue: V) => void;
+  value?: V;
+  hidden?: boolean;
 }
 
 export const More = () => {
@@ -23,25 +32,56 @@ export const More = () => {
   const handleChangeHapticFeedback = (_hapticFeedbackEnabled: boolean) =>
     dispatch(setHapticFeedbackEnabledAction(_hapticFeedbackEnabled))
 
-  const hapticFeedbackEnabled = useSelector(hapticFeedbackEnabledSelector)
+  const handleChangeNativeSearchEngine = (_nativeSearchEngineEnabled: boolean) =>
+    dispatch(setNativeSearchEngineEnabledAction(_nativeSearchEngineEnabled))
 
-  const renderItem = ({ item: { title, onChange } }: { item: MoreOption<any> }) => (
+  const handleClearSearchHistory = () => {
+    clearSearchHistoryAlert(() => {
+      Storage.set(STORAGE_KEY.SEARCH_RESULT, JSON.stringify([]))
+    })
+  }
+
+  const hapticFeedbackEnabled = useSelector(hapticFeedbackEnabledSelector)
+  const nativeSearchEngineEnabled = useSelector(nativeSearchEngineEnabledSelector)
+
+  const renderItem = ({ item: { title, value, hidden, onChange } }: { item: MoreOption<any> }) => hidden ? (
+    null
+  ) : value ? (
     <ListedOption title={title}>
-      <CustomSwitch defaultValue={!!hapticFeedbackEnabled} onValueChange={onChange} />
+      <CustomSwitch defaultValue={value} onValueChange={onChange} />
+    </ListedOption>
+  ) : (
+    <ListedOption title={title}>
+      <TouchableOpacity hitSlop={{ top: 10, right: 10, bottom: 10, left: 10 }} onPress={onChange}>
+        <MaterialCommunityIcons name="delete" color={COLOR.BLACK} size={26} />
+      </TouchableOpacity>
     </ListedOption>
   )
 
-  const isPending = R.any(R.isNil, [hapticFeedbackEnabled])
+  const isPending = R.any(R.isNil, [hapticFeedbackEnabled, nativeSearchEngineEnabled])
 
-  const options: [ MoreOption<boolean> ] | [] = isPending ? [] : [
-    {
-      title: 'Haptic feedback',
-      onChange: handleChangeHapticFeedback,
-    },
-  ]
+  const getOptions: () => [ MoreOption<boolean>, MoreOption<boolean>, MoreOption<any> ] | [] =
+    React.useCallback(() => isPending ? [] : [
+      {
+        title: 'Haptic feedback',
+        value: !!hapticFeedbackEnabled,
+        onChange: handleChangeHapticFeedback,
+      },
+      {
+        title: 'Native search Engine (beta)',
+        value: !!nativeSearchEngineEnabled,
+        onChange: handleChangeNativeSearchEngine,
+      },
+      {
+        title: 'Clear search history',
+        onChange: handleClearSearchHistory,
+        hidden: Platform.OS === 'ios',
+      }
+    ], [ hapticFeedbackEnabled, nativeSearchEngineEnabled, isPending ])
 
   return (
     <MoreContainer topInset={topInset}>
+      {nativeSearchEngineEnabled}
       <FlatList
         renderItem={renderItem}
         ListEmptyComponent={(
@@ -49,7 +89,8 @@ export const More = () => {
             <ActivityIndicator size="large" />
           </MoreActivityIndicatorWrapper>
         )}
-        data={options}
+        extraData={nativeSearchEngineEnabled}
+        data={getOptions()}
       />
 
       <ListedOption titleSize={TEXT_SIZE.XS} title={`App version: ${app.version}`} />