+++ src/native-db/hooks/use-native-sb-events.hook.ts
@@ -1,23 +1,28 @@
 import * as React from 'react'
 import { DeviceEventEmitter, NativeEventEmitter, NativeModules, Platform } from 'react-native'
 
-export const useNativeDBEvents = () => {
+export const useNativeDBEvents = (
+  resultCallback: (result: string[]) => void,
+) => {
   React.useEffect(() => {
-    const eventEmitter = new NativeEventEmitter(NativeModules.EventEmitter)
+    const eventEmitter = Platform.OS === 'android' ? null : new NativeEventEmitter(NativeModules.EventEmitter) as any
 
     if (Platform.OS === 'android') {
       DeviceEventEmitter.addListener('findPossibleWordsResult', (result: string) => {
-        console.log('FROM NATIVE: findPossibleWordsResult: ', JSON.parse(result)[0])
+        resultCallback(JSON.parse(result))
       })
     } else {
       eventEmitter.addListener('findPossibleWordsResult', (result: string) => {
-        console.log('FROM NATIVE: findPossibleWordsResult: ', result)
+        console.log('FROM NATIVE iOS: findPossibleWordsResult: ', result)
       })
     }
 
     return () => {
-      eventEmitter.removeAllListeners(NativeModules.EventEmitter)
-      DeviceEventEmitter.removeAllListeners()
+      if (Platform.OS === 'ios') {
+        eventEmitter.removeAllListeners(NativeModules.EventEmitter)
+      }
+
+      DeviceEventEmitter.removeAllListeners('findPossibleWordsResult')
     }
   }, [])
 }