+++ src/native-db/native-db.constants.ts
@@ -1,0 +1,1 @@
+export const NATIVE_DB_TAG = 'NATIVE_DB'