+++ src/native-db/native-db.models.ts
@@ -1,6 +1,9 @@
 import { NativeModulesStatic } from 'react-native'
 
 export interface NativeDB {
-  findPossibleWords: (allWords: string[]) => string[];
+  findPossibleWords: (
+    allWords: string[],
+    selectedLetters: string[],
+  ) => string[];
   _nativeModule: NativeModulesStatic;
 }