+++ src/native-db/native-db.ts
@@ -4,8 +4,15 @@ import { NativeDB } from './native-db.models'
 const _nativeModule = NativeModules.DBModule
 
 export const DB: NativeDB = {
-  findPossibleWords: (allWords: string[]): string[] => {
-    _nativeModule.findPossibleWords(JSON.stringify(allWords))
+  findPossibleWords: (
+    allWords: string[],
+    selectedLetters: string[],
+  ): string[] => {
+    _nativeModule.findPossibleWords(
+      JSON.stringify(allWords),
+      JSON.stringify(selectedLetters),
+    )
+
     return []
   },
   _nativeModule,