+++ src/settings/store/settings.selectors.ts
@@ -2,3 +2,6 @@ import { RootState } from '../../store/store'
 
 export const hapticFeedbackEnabledSelector = (state: RootState) =>
   state?.settings?.hapticFeedbackEnabled ?? true
+
+export const nativeSearchEngineEnabledSelector = (state: RootState) =>
+  state?.settings?.nativeSearchEngineEnabled ?? false